MB-40583 Explode secExprs in client for flattened array index
For flattened array index, explode the secExprs string. E.g.,
for the index:
create index idx on default(org,
	DISTINCT ARRAY FLATTEN_KEYS(v.name, v.age DESC, v.year) for v in dob end, email)
The secExprs array will contain
	[`org`, `DISTINCT ARRAY FLATTEN_KEYS(v.name, v.age DESC, v.year) for v in dob end`,
	  `DISTINCT ARRAY FLATTEN_KEYS(v.name, v.age DESC, v.year) for v in dob end`,
	  `DISTINCT ARRAY FLATTEN_KEYS(v.name, v.age DESC, v.year) for v in dob end`,
	  `email`]

The number of times the "FLATTEN_KEYS" expression is repated inside
the secExprs is proportional to the number of fields in the
"FLATTEN_KEYS" expression.

The `secExprs` slice is exploded for two reasons:
a. To match the DESC attributes for flattened array expressions.
   In the above example, the desc slice is constructed as
   [false, false, true, false, false]. The length of desc slice
   and len(secExprs) are made equal
b. For scans which depend on key positions (like group, groupAggrs),
   match the key position in scans with the key positions in
   secExprs. E.g., keyPos 4 in scans will map to keyPos 4 in secExprs
   i.e. `email`

Change-Id: I5392587a32e66f9f7b40439937c9ac84b591179e