MB-40583 Pass unexploded version of secExprs to projector
For a flattened array index, secExprs field in index definition
is exploded with the flattened array expression. The number of
time flattened array expression is repeated is proportional to
the number of fields in the flattened array.

However, when evaluating a document against secExprs, it is
sufficient to evaluate the document only once against the
flattened array expression. The other repeated versions of
flattened array expression in secExprs can be ignored.

Hence, passing on the original list of secExprs to projector
from kv_sender

Change-Id: I0b488c763c55372a4e9e6c000c29e5d61a806e3b