MB-47632: [BP 7.0.2 MB-46883] - [System Test] Rebalance failed with error "Index bucket7.sg_roleAccess_x1 already exists"
Backport of MB-46883 from 7.1 to 7.0.2

When node is failedover another index with same name as that of index on the failed node can be created, when the failed over node is added back to cluster, cluster now has duplicate index, these may or may not be equivalent indexes but they have same name. This sometimes causes rebalance to fail with error index already exists. The fix addresses the issue by detecting the duplicate indexes and deleting the duplicate index if they are equivalant. In case of duplicate indexes being non-equivalant system behaves as it used to behave earlier which means there is a chance of rebalance failure with index already exists.

Change-Id: Iecade401029d2aafbd05ee2db5cfcd13ddc01dec