MB-37671 : Allow specifying secure port in query
Currently we dont allow user to specify the secure port numbers in the
query for Move or Create index. After this change we should be able to
give secure ports. When Encryption is enabled query can contain secure
ports. If Encryption level is Strict i.e. we disable NonSSLPorts query
should only contain secure ports non-secure ports will not be allowed in
the query. This patch will change the output of getIndexStatus too the
node names will contain secure ports when the encryption is enabled.

Change-Id: If08deacd1e35cd836ca07511a32ff2ee9713b962