MB-47666 : Increase polling interval in pollForDeletedBuckets
This polling is calling bucket/<bucketName> which is a costly call.
Increasing the polling interval to reduce load on ns_server. This is
used to print numDocsPending stat in projector. This is used to cleanup
seqno reader when the buckets get deleted so increasing the interval
should not have high impact.

Change-Id: I9ff5cb6dfde57b60161b5ae3abe948ea656c613a