MB-47827 Validate the presence of existing tokens in MoveIndexTokenPath
BP of MB-47775 to Cheshire-Cat

Before upserting the MoveIndexTokenPath, check for the presence
of existing tokens in metakv.

If a token exists and it is different than the caller's version,
report error. If a token does not exist and an update is required,
do not update metakv. In other cases, update the token.

Change-Id: I324ec72bf68679512b178076cffba8aab270d80b
(cherry picked from commit 37f2b3e07377ab378b4857b86c638ca1803e538a)