MB-48157 Reduce idle time CPU utilisation of projector
Idle time CPU utilisation is high because the endpoint::run
method wakes every 1ms for flushing the bucker. The ticker
interval has been reduced to 1ms to achieve low session
consistent scan latencies with less KV ops/sec rate.

With this patch, the flushTick is moved back to 25ms. When
a mutation is to be processed, we flush it if there are more
than 100 mutations in the queue or the time since last flush
is greater than 1ms - so, scan latencies should not be impacted.

Change-Id: I67718e6f1b971e46d9137145e6fb5996ca3771d5