MB-48017 Reduce idle time CPU utilisation of projector
BP of MB-48157 to 7.0.2

Idle time CPU utilisation is high because the endpoint::run
method wakes every 1ms for flushing the bucker. The ticker
interval has been reduced to 1ms to achieve low session
consistent scan latencies with less KV ops/sec rate.

With this patch, the flushTick is moved back to 25ms. When
a mutation is to be processed, we flush it if there are more
than 100 mutations in the queue or the time since last flush
is greater than 1ms - so, scan latencies should not be impacted.

Change-Id: I67718e6f1b971e46d9137145e6fb5996ca3771d5
(cherry picked from commit 2c566914c946872e641014c5faece10eb3c45f1d)