MB-47600 Treat NaN rebalance progress as 0.1
When indexer participates in rebalance, the transfer tokens
will not be computed until indexer finishes planning. During
this time, indexer reports progress as NaN to ns_server in
the output of GetTaskList. Go's JSON encoder does not know
how to encode NaN and the rpc code ignores these encoding
errors. So, the response to GetTaskList will not be sent to
ns_server and this can cause rebalance failures

Instead of failing rebalance, indexer would now start sending
progress as 0.1 even if the progress computed is NaN.

Change-Id: Ibc7c3fce8ec37e1c54015a0cc60f4619d1d28a8b