MB-48300 Retry shutdown with local kvaddrs incase of node rename
In case where a node gets added with localhost address first
and then ns_server changes the IP address to a public/private
IP, then the feed.nodeFeeds book-keeping becomes inconsistent.

E.g., when a node is added with 127.0.0.1 as IP address and
streams are opened, the key to nodeFeeds would be 127.0.0.1
When ns_server changes the IP address, then feed.bucket.getMasterNode
would return the changed IP address. This new IP address will not
be a part of nodeFeeds book-keeping and this call will fail.

Currently, shutdownVBuckets code path does not handle errors. So,
the memcached.ErrorInvalidFeed error gets silently ignored and
the vbucket will never be shutdown. This will lead to indexer endup
in a repair loop. To fix this, we check if the master node's IP address
is same as the local IP address. If yes, we go ahead and shutdown the
stream by using the local kvaddress

Change-Id: I1f02cd863f707ad04bbec4d47262f28878bc70ff