MB-48600 Restart inactive MAINT_STREAM during stream merge
If MAINT_STREAM is deleted after TK_INIT_BUILD_DONE_ACK message
is processed, then we need to restart the MAINT_STREAM. The logic
to restart MAINT_STREAM is present in handleMergeStreamAck() method.
However, this logic gets skipped as we check for the INIT_STREAM
state and skip this logic if INIT_STREAM is in state STREAM_INACTIVE.

As we enter handleMergeStreamAck() method under the pre-condition
that INIT_STREAM state is set to STREAM_INACTIVE, the check in
handleMergeStreamAck() is no longer required.

Change-Id: I23af21ff70cc3f5868c4f4447152599db0e942d2