MB-48580: Magma GSI index test hit "panic: runtime error: invalid memory address or nil pointer dereference"
RCA: when indexer is in warmup and the first index's recovery from disk is still in progress,
if there is a stats request made to storage manager, before the recovery of first index this panic does happen as the stats is nil.
Storage Manager's "stats" object gets initialized when indexer sends the message to update IndexInstMap after the recovery of first index in indexer warmup phase.

Solution: Initialize the "stats" object of storage manager when initializing the storage manager itself.
Change-Id: I982c63faa6a2e3febeb857a071344bfea7bb088f