MB-54690 Do not cleanup index data when flush is in progress
BP of MB-48754 to 7.0.5

If collection drop happens while flush is in progress,
then indexer will wait for flush to be done. After flush,
the indexes in dropped keyspaces will be removed.

However, with the fix for MB-46054, timekeeper can send
a INDEXER_KEYSPACE_NOT_FOUND message when a collection
is dropped while stream request for a vbucket is in
progress. In such cases, indexer will go-ahead and
clean-up index data while flush is in progress.

The fix for this issue is to avoid index clean-up if
flush is in progress

Change-Id: Ie715a1a0149295e9007cafcec4dbd4e20e0437dd
(cherry picked from commit 678a5181bf3e26f928cf8c974948b220f4ab9d5d)