MB-48869 Add missing nil check for GetLocalIndexMetadataFromCache return
Trivial fix for regression:

http://review.couchbase.org/c/indexing/+/162484 eliminated the error
return from GetLocalIndexMetadataFromCache(). One call of this needs to
check if the returned metadata is nil. Previously it checked if the
error was nil but this check was removed in the aforementioned change,
but it should have added a replacement nil check on the metadata value.

Change-Id: I1fa0170d15db7b820c4a6cec3287733a4557cd2b