MB-44738 Part 3 (7.1.0 1469) Autofailover for Index Service feature
Part 3 updates AutofailoverServiceManager and RebalanceServiceManager
(fka ServiceMgr) to match the final ns_server API for the former and
refactor to meet the new requirement that ns_server RPC registration
must be done with a single object that implements both interfaces, as it
does not support independent registration. This is accomplished by
adding a new parent class, MasterServiceManager, that contains instances
of the other two service manager classes and delegates all calls to
those. MasterServiceManager is now the class registered with ns_server.

Change-Id: I61baa9bae4fcc40074add5b20ada8df803e4f436