MB-48885 Smart Batching: properly handle globalTopology == nil
Index move case will not run the planner, so globalTopology may be nil.
Updated initNodeLoads() to just return in this case instead of trying to
deference the nil pointer. Since the number of indexes on each node also
came from globalTopology and then was updated based on transfer tokens
published rather than from the stats, I deleted all the index counting
code rather than try to update it to work another way. It was used as a
minor component of deciding if a node is highly loaded or not but is
really an edge case, and as soon as the stats are retrieved we have
better information on load levels, so it is better to simplify this out.

Change-Id: I077f2df1c262b68d4f4ca833240f1f73cca47f98