MB-48980 Smart Batching: Properly handle topologyChange == nil
Rebalancer.topologyChange is nil in the Move Index case, which caused a
panic when Smart Batching deferenced it to learn the number of
KeepNodes. Fixed to assume 1 if topologyChange == nil.

Failover case is done mostly in RebalanceServiceManager so does not
execute much Rebalancer code and did not trigger this bug. It does have
topologyChange in RebalanceServiceManager but was passing nil to
NewRebalancer anyway. I changed it to pass the topologyChange pointer
for better code hygiene even though it is not currently used.


Change-Id: If3a957f019e58611d528625ffbd238bdd482ca9e