MB-45484 Incrementally update bootstrap stats
With the current logic, the entire payload of bootstrap
stats is marshalled and sent to lifecycle manager. With
10000 indexes on a node, the time to marshall and
unmarshal the stats for 10000 indexes takes long time and
this can delay indexer getting ready.

To avoid this, instead of sending the entire stats payload,
only the stats of the index which got recovered are sent to
lifecycle manager. The client related stats are updated by
timekeeper which can only process the stats for the instances
once storage warm-up is done. This means that all client stats
for each index will contain the values initialised during
bootstrap - until storage warm-up is done. So, it is safe to
send only the stats for the index.

Change-Id: I9c5642df3988423ecffaaa9dcefee4031f40b2cd