MB-48730 (7.1.0 1469) Rework DDLServiceMgr post-Rebalance processing
When a Rebalance finishes it reports to ns_server that it is done, but
it was still doing background cleanups under go notifyRebalanceDone(),
one part of which would get and hold the ddl_service_manager.go global
gDDLServiceMgrLck mutex while doing possibly extensive cleanup,
including deleting old DDL tokens from metakv. This was the root cause
of the current bug as the following Rebalance timed out attemping to
lock this same mutex in stopDDLProcessing()'s call to
getDDLServiceMgr().

The async cleanup code had become disorganized after years of
miscellaneous updates. This fix reorganizes it and eliminates the need
for it to hold the problem mutex. It also consolidates token cleanup
under a single goroutine and changes write locks to read locks in a few
places, including in the related file sched_index_creator.go, and it
allows DDL to begin processing again sooner without waiting for async
cleanups. DDL processing was held blocked in the old code because the
token cleanups did not used to check the allowDDL flag and just forged
ahead, but now they check it and token cleanup is no longer done in the
foreground of Rebalance cleanup after this fix either.

Change-Id: I64dd8b87363c5ae4cb6871697d8fccebe235275a