MB-49083 (7.1.0 1542) Do not add nil stats entry to respMap on failure
stats.go parallelStatsRestCall was incorrectly adding a nil stats entry
to its respMap return map if stats failed to retrieve for a node. An
entry should only be added to the errMap return map in this case. The
presence of nil entries in respMap caused Rebalancer.getNodeIndexerStats
to incorrectly call addStats with a nil stats pointer, triggering the
panic for nil pointer deference.

Change-Id: Ia4cb08e9a46ce3113e542e0926280e0d3d0ab88e