MB-49080 Populate missing values for all keys for flatten array index
If array is non-leading key and is empty, then "missing" has to
be populated for all keys in flatten array expression. Otherwise,
indexer side ExplodeArray4 logic would break. Similarly, "empty"
and "null" cases are being handled.

Also, if flattened array is leading key in index and if the first
value of flattened expression is missing, the document is filtered
from indexing as we do not support index with missing leading key

Change-Id: Ia0615f0271d5d4157c0544ec63b27cd27bcf6a1c