MB-48273: Index creation with empty magma bucket is stuck forever.
This is not a magma specific issue. When there are no mutations and DCP_SEQNO_ADVANCE message is missed but MAINT_STREAM has high sequence number (e.g due to creation of collections)
INIT_STREAM for collection is not flushed and hence the initFlushTs is nil. This causes stream merge not to proceed.

Fix: initialize the initFlushTs sequence numbers to all zero vector and try the merge.
Change-Id: I0ce09d7aa8f539c400e4e608703f26375e18b021