MB-48299 (7.1.0 1558) Add IPv6 support to utils.go GetLocalIP()
utils.go GetLocalIP() was missing six lines of code that are needed for
IPv6 support. The fix is to copy the body of dcp/upr.go getLocalIP()
which has these lines. (There are two copies of this function because
upr.go cannot include the common package as it creates an import
dependency cycle.)

Change-Id: I579480bbe5c9678349c0cfec90c66e918f1619a7