MB-44775: All array index show wrong item indexed count in UI
Storage layer tracks items_count as number of index entries in storage layer, this works for DISTINCT array index correctly BUT for the ALL array index storage will store only a single entry and any repetations are encoded as the count inside the entry, due to this the items_count would not account for count of repetations.
FIX: maintain seperate counter at GSI layer which will account for repetations correctly, this new stat is stored along with snapshot. Available for MOI and Plasma indexes created after all indexers are at 7.1 or higher.

Change-Id: Ia2e2dd94a32b10ad4257f26aa4135e29b7e18847