MB-46251 Handle errors in shutdownVBuckets code path
ShutdownVbuckets is a best effort call from indexer. So far,
any errors during this call are not handled by cleaning-up the
keyspace. This can become a problem in some cases when projector
fails to shutdown the vbucket, indexer attempts to restart but
restart will not be successful as the stream is considered
active by projector. This will result in indexer book-keeping
becoming corrupt and sometimes not recovering from restart
loop.

In this patch, we attempt to clean-up keyspace if any errors
were observed before attempting to shutdown the vbuckets. This
will make sure that indexer will not endup in an endless restart
loop

Change-Id: I25c00b8b920b6d7c7d2f46172d3ea6a95f878c88