MB-49356 Part 1 (7.1.0 1695): Autofailover: IsSafe() more logging
Autofailover IsSafe(): Log both the cached and current set of index
statuses as JSON when returning unsafe. This is done in a goroutine as
it must do scatter-gather to get the current index statuses, which can
take a long time. Also, only log this if it has been at least 5 minutes
since the prior logging as the index status JSON can be huge.

Change-Id: I27007160ca40e576b62a2bb0271014815a7f1093