MB-49838 Revert back to using collection seqnos for session consistent scans
Using bucket_seqnos for session consistent scans will help
reduce the init_latency of the scan but it will increase
the scan_wait latency as each scan has to wait for the
snapshot to be ready even though no mutations have happened
on the collection. This is against the multi-tenency nature
of collections. Hence, reverting back to using collection
seqnos.

The existing setting can be used in specific cases where the
scan_init latency is observed to be very high. We can then
enable using bucket_seqnos on a need basis

Change-Id: Ifb869656291411f267f10f0aa616fee279b8ec07