MB-49079 (7.1.0 1861) Rebalance: Omit inconsistent indexes from Planner
This is a work-around for when Rebalance encounters MB-47874, since that
MB is still under analysis. Both of these came from CBSE-10499.

In the Rebalance case only, if Planner finds indexes with inconsistent
numbers of replicas in their metadata, these indexes will be removed
from the layout so Planner and subsequent Rebalance can proceed instead
of failing. This makes those indexes invisible to Planner, so the
resultant plan may be far from optimal, but this is deemed better than
failing Rebalance.

Change-Id: Ib23233d776c5675c04586e0709854fe8ae156eaf