MB-50058 Skip vbuuid check if target seqno is '0' during snapshot merge
When there are less number of documents in the bucket and some vbuckets
have no documents, then it is possible for timekeeper to generate a disk
snapshot even if some stream begin messages are pending to be processed.
The disk snapshot in that case would not contain vbuuid's of all vbuckets.
When merge is attempted with such snapshot, this method would return "false"
as some Vbuuid's would be empty in source timestamp. So, the caller of this
method would panic and indexer would restart.

One way to fix the issue is to wait for all stream beings and only then
generate disk snaphosts. However, this logic will effect all index build
cases. Since the issue is only when vbuckets have "0" seqno, this method
would skip the vbuuid check if the vbucket seqno is zero.

Change-Id: Ie220567322d09cab723ed85aa768b380949b97db