MB-50165 Proceed with merge when initFlushTs goes past maintFlushTs
For INIT_STREAM to be merged to MAINT_STREAM, following conditions
are to be satisfied:

a. initFlushTs >= minMergeTs && initFlushTs >= maintFlushTs
b. initFlushTs >= minMergeTs && initFlushTs < maintFlushTs
   initFlushTs >= collection seqnos &&
   maintFlushTs >= bucket seqnos
c. initFlushTs < minMergeTs && initFlushTs >= collection seqnos

In the third case, if timekeeper decides to merge and if
initFlushTs >= maintFlushTs && maintFlushTs < minMergeTs,
then timekeeper will disable flush and force a snapshot
with maintFlushTs.

After the snapshot is processed, the merge check happens
again with initFlushTs (which is now equal to maintFlushTs).
But, as maintFlushTs < minMergeTs, and if the collection seqnos
advance beyond maintFlushTs by this time, merge will never happen
and timekeeper continues in loop and this will stall snapshotting.

In this patch, we address the issue by comparing initFlushTs and
maintFlushTs when snapshot type is FORCE_COMMIT_MERGE and
maintFlushTs < minMergeTs. If initFlushTs >= maintFlushTs and
vbuuid's match, the stream will be merged

Change-Id: I335c9574ae4ef80300c13470265a08b2470e2441