MB-50273 (7.1.0 2004) Autofailover: Do not consider scheduled indexes
request_handler.go getCachedIndexTopology() should not use
ClusterInfoCache but its construction of status information for
scheduled indexes did use this (see the MB for call chain). This
could not only take a long time, but also this method was not holding
the cinfo's mutex so if something else changed it while it was being
read here, it could cause request handler to crash.

Autofailover should not consider scheduled indexes so the fix is to
remove this computation. Also add an optional "omitScheduled" flag
(default false) to getIndexStatus() so Autofailover isSafeLogDetails()
can set omitScheduled=true so the non-cached statuses it logs will
match the cached ones in omitting scheduled indexes (otherwise
debugging incorrect IsSafe() decisions would be more difficult).

Change-Id: I84750492fdd8ab2a082ed58a3b528ad02282909d