MB-50006 Use the same snapContainer even after rollback
The current code is built on an assumption that an instance
is tied to a single snapshot container for its entire lifetime.
This assumption is broken in updateIndexSnapMapForIndex code
when this method is called after rollback.

This patch fixes the issue by using the same snapshot container
as before rollback to zero

Change-Id: I3069c09bef12d8a13c91e2a8a60ab3f1fcd1590e