MB-49153 avoid sessionId collision on indexer restart
If indexer gets restarted soon after starting, it can reuse the same
sessionId as it always gets computed from 1. This can lead to
stale mutations from projector coming in with a valid sessionId
and get processed.

The current fix chooses a random number between 1 to 1000 for the first
sessionId. This should be sufficient as frequent indexer restarts are
not very common. The other option can be to persist the last used
sessionId. The current solution should be sufficient and the extra
persisted state management seems to be an overkill for now.

Change-Id: If086b3078f35da264ebae4199aa64530574e93ea