MB-46610 (7.1.0 2037) Fix more TCP connection leaks
To avoid leaking TCP connections (until their 2 min timeout expires),
the caller must read the entire body to the end AND close the body
(verified in many places on the web), otherwise that connection will not
be reusable until it times out, because Go still holds the buffer until
the timeout in case any other goroutines are reading it EVEN IF
res.Body.Close() is called.

The hard-to-find meme for avoiding this problem is:

  defer res.Body.Close()
  defer io.Copy(ioutil.Discard, res.Body)

The io.Copy to ioutil.Discard will read any previously unread portion of
the body, which will change Go's internal state tracking to show the
entire body has been consumed, so Go will no longer hold the buffer for
two minutes, and the Close() call will thus clean up and make the TCP
connection reusable immediately.

Change-Id: I50b1c8f344b0c17f2b857474bed4700395a93cda