MB-50431 do not snap align FORCE_COMMIT_MERGE timestamp
The index(es) being merged to MAINT_STREAM will become part of the
MAINT_STREAM processing and any required snap alignment to the timestamp
will happen as part of MAINT_STREAM processing.

Also, ensureMonotonicTs check is more of a safety check. As no flush is
invoked in case of FORCE_COMMIT_MERGE timestamp, it shouldn't be
required.

Change-Id: I5d883957d5ce56e3de0fcc50d6e8fb6598de0891