MB-50659: Enable dataport auth
Backport MB-50322 to 7.0.4

The indexer sends stream requests to the projector, which
causes projector to start endpoint client. The endpoint
client is created synchronously to any request made by the
indexer. Indexer running on 7.1+ version will always send
needsAuth = true. Projector can send auth packet in the
mixed mode cluster, as that indexer is known to support
the auth.

Server will always process the auth packet as long as
it is upgraded.

Change-Id: Ic2fe2b0a4e93922d2516b782b8cc02bad838c914