MB-50397 Share the request/stopCh for OSO build recovery
When a stream request is made to projector, indexer will update
the streamKeyspaceIdCurrRequest map and then process the request.
If the stream request fails due to an error on projector side,
indexer would try initiate recovery. Recovery would proceed only
when the requestCh in streamKeyspaceIdCurrRequest and the requestCh
passed to handleInitPrepRecovery method are the same. This is to
avoid multiple recoveries from happening simultaneously.

For OSO builds, if recovery were to be initiated, this channel is
not being passed while the map streamKeyspaceIdCurrRequest is set.
This causes a mismatch and recovery would never proceed. This patch
fixes the issue by passing the requestCh to handleInitPrepRecovery
for OSO builds as well

Change-Id: I30e25d021be9709cde77f357fe91f7028053599a