MB-50664: Indexes stuck at state building in travel-sample bucket
In checkInitialBuildDone if buildTs is zero we mark the initBuildDone as true and proceed with TK_INIT_BUILD_DONE.
This is not sufficent in OSO mode as flushTs can have open OSO snapshots.
When MAINT_STREAM is not yet started this flushTs with its openOSOSnapshot flag gets set on the MAINT_STREAM. Later in checkInitStreamReadyToMerge it can get set back to INIT_STREAM.
Due to this even when actual ts does not have any OSOSnapshot markers the openOSOSnapshot flag on ts is true which prevents the INIT_STREAM from merging into MAINT_STREAM
with error - has open OSO snapshot.

Fix: In the cehckInitialBuildDone for buildTs is all zero case additionally check flushTs is snapAligned and for OSO mode it should not have any OpenOSOSnap.

Change-Id: I7c1e53d5fa21ea5e1fe3f9311e3f1ed3b0d27ff3