MB-50770 (7.1.0 2179) Reduce retries under waitForScheduledIndex()
cluster_info.go FetchForBucket() is only called by metadata_provider.go
checkValidKeyspace(), a private function inside waitForScheduledIndex().
This had a 300x RetryHelper inside FetchForBucket(), but the caller has
its own retry loop that terminates as soon as an ending condition is
detected. One expected ending condition is that the bucket, scope, or
collection has been dropped, so this should not retry hundreds of times.
Doing so floods query.log with retry messages from children.

The fix is to change the RetryHelper from 300 to 5 retries, as well as
to change the children (and related functions while we're at it) that
have their own internal loops that do 5 retries to only log one message
for all retries instead of one per retry.

Change-Id: I3419e8f72d363dcad7ae0ac297a8c05a8ea35057