MB-50958 Return NIL collection or scope ID's when bucket is not found
The current cinfo lite infra will return ErrBucketNotFound when
retrieving the collection info if bucket is not existing in the
cluster. This will lead to subsequent operations being skipped.
(E.g., monitorKeyspace will try to getCollectionID but as
GetCollectionInfoProvider returns ErrBucketNotFound, it returns
without cleaning up indexes)

This patch addresses the problem by returning COLLECTION_ID_NIL
and SCOPE_ID_NIL when bucket is not found

Change-Id: Iba4b69d5cd6a537a2b6c4798c678527667879641