MB-51096 Do not instantiate cinfo client in internal version monitor
CInfoProvider will either instantiate cluster info client or lighter
version of cinfo client. In cases where the code is expected to do
manual fetches like FetchNodesAndSvsInfo, then it can be a problem
as the following race condition can kick in:

a. The cluster info client instantiated will start listening on all
   bucket streaming endpoints
b. Caller calls FetchNodesAndSvsData() on the cluster info cache of
   the client which will overwrite the pool information as we do
   GetPoolWithoutRefresh() - So, it will have a nil manifest
c. The cluster info client will see a CollectionManifestChangeNotification
   and then tries to update the manifest
d. As manifest is nil - due to GetPoolWithoutRefresh() indexer will panic

It is not expected to do manual fetch on the cinfo object of client.
To address the panic, this patch will instantiate a direct cluster info
cache object instead of going through ClusterInfoProvider

Change-Id: Ic16d2643c06f3197e90830bf452af5b7ecb9752b