MB-51029: Indexes Replica on Failed over node is not getting build when the node is added back with recovery
When an indexer node is in failedover state its indexes are not available to check state, token cleanup performed by DDLServiceMgr ends up cleaning up the token even if one of the replica which is on failedover node is not yet built. Once the build token is gone, later even if node is added back with full recovery the index would not be built and remain in created state. Skip token cleanup if there are failed indexer nodes in cluster.

Change-Id: I441efef9166e58ec177850ab54ba470c435c1bf0