MB-50943 (7.1.0 2284) Fix boot init of Index Service node topology
During Indexer bootstrap rebalance_service_manager.go updateNodeList()
was called asynchronously, with no retries, to initialize the list of
known Index Service nodes via a REST scatter-gather to all known Index
nodes according to ClusterInfoCache. This had two problems:

1. The inital REST call to localhost to the handler for this scatter-
   gather would fail because the cbauth database has not finished
   booting. (This failed on every boot in a dev environment. In
   production it usually did not hit this failure but sometimes did.)

This is fixed by retrying for 60 seconds. In a dev env cbauth takes a
small number of seconds to finish booting. Then this call succeeds.

2. ClusterInfoCache does not list any Index nodes that are not fully
   bootstrapped as being known Index Service nodes, so the scatter never
   went to the current node, as it was in the midst of booting, nor any
   other Index node that was not yet fully booted, so the list of nodes
   returned could be a subset of the true known list, including an empty
   list (which was the case for the first node to start).

This is fixed by eliminating the scatter-gather and use of CIC and
instead getting the last known node topology from the local
requestHandlerCache, which is updated and persisted every five seconds
during normal operation and reloaded from disk at boot. Thus except in
extremely narrow edge cases this will contain the correct set of nodes.
(If it is incorrect a Rebalance will still fix it, but this will be very
rare.) This is the same source of info Autofailover uses to decide
failover safety.

Also, the original scatter-gather call used was designed for something
else and returned the full index topology for all indexes on all known
Index nodes. The caller then threw everything away but the NodeUUIDs of
the Index nodes. The new call returns only the NodeUUIDs.

Change-Id: Ieb53566df12d35978ac32c234eb18418f457eeec