MB-51096 Keep reader and response body in sync
If a new reader is initialised everytime, it always
reads from 0th location of response body. If we assume
res.Body is like a circular buffer, then the 0th entry
might actually point to a stale data which can lead to
unmarshal errors. This patch fixes the issue by keeping
reader and res.Body in sync instead of initialising a
new reader everytime

Change-Id: I39d51db2362dfc64f4225508d3050cbd97751959