MB-51025: Avoid data cleanup for tombstone inst if data is valid
During rebalance, if indexer restarts while partition merge was
in progress, then it is possible to end up with a conflicting
proxy instances. A proxy instance created by merge partition
operation can have a valid recovery state and valid data.

During recovery, if such conflicting instance is ignored, the
tombstone instance cleanup can lead to cleaning up of the valid
index data, which can leave the index in an inconsistent state.

The fix avoids cleanup of the data if tombstone instance has
a valid conflicting instance.

Change-Id: I1c3f9985f51ce5694d4d08c2a47d25e5526de92f