MB-50660: Avoid log flooding during kv-rebalance before starting indexer service on node addition - during upgrade
Backport MB-51274 to 7.0.4

Log flooding happens due to difference in our interpretation of what is a active service and what is configured service.
Querying pools/default endpoint only confirms the cluster membership of node as active, which only means service is configured on that node but it does not always mean service is active on that node. To get list of active nodes for a service we need to query pools/default/nodeServices endpoint.
In the internalVersionChecker codepath GetActiveIndexerNodes() was being used to get list of activeNodes but its return value comes from pools/default endpoint which is list of configired services and not necessarily active services.
Fix:- renamed GetNodesByServiceType to GetNodeIdsByServiceType as the function returns index of nodes and not the actual nodes, added new GetNodesByServiceType which retruns list of active nodes for service which comes from pools/default/nodeServices endpoint. GetNodesByServiceType is now being used by internalVersionChecker code path to get list of active nodes instead of GetActiveIndexerNodes/GetActiveKvNodes/GetActiveQueryNodes.

Change-Id: I63e505c7cbd6a3b8c78f18cf287132352656eb61