MB-51363 Force INMEM_SNAP generation after recovery
After indexer recovers, it is possible for some instances to have
incorrect vbuuid's. This can happen if indexer restarts in the
middle of disk snapshot generation after vbuuid's change. In that
case, there will be a mismatch in vbuuid's between the timestamp
of an instance at storage manager and timekeeper. If timekeeper
were to pick the old vbuuid's, KV would ask indexer to restart
streams using latest vbuuid's. However, if timekeeper picks latest
vbuuid's, then some instances in storage manger will be left with
older vbuuid's. To address those cases, timekeeper will force a
INMEM_SNAP generation so that all the instances belonging to the
keyspaceId will have same vbuuid's

Change-Id: I0c931e9e34119cd50956959b3218de3d74b04b42