MB-51330 handle recovery of INIT_STREAM in Catchup phase
After INIT_STREAM has entered the catchup phase, if recovery starts, it
will clear the timekeeper bookkeeping and add indexes back to it later.
But this leads to minMergeTs being lost in timekeeper bookkeeping.

With this fix, indexer will store the minMergeTs during catchup phase of
INIT_STREAM, which can be used by OPEN_STREAM message from indexer to
timekeeper to set the minMergeTs during recovery.

Change-Id: Ie0b81e5601273a8514597f1a7f9ab6ddaed0d420