MB-51570: Indexer crash with error "panic: Fail to find real index instance 3055985754217342362"
RCA:
1. 2 partitions of the same index are being moved to a single node from different destinations in the same rebalance batch.
2. During the initial build of these instance (1 proxy and 1 real), the rebalance gets cancelled or fail leading to cleanup of in progress index builds.
3. The first token to be picked up for cleanup is the one with real instanceId.
4. When the second token gets processed, before its drop index can get processed by indexer, its build finishes.
In this case there is a proxy instance which moved to Catchup but there is no real instance. This violates the constraint check in Indexer::processBuildDoneCatchup and indexer panics.

Fix: During rebalance cleanupTransferTokens identify 2 or more partitions of the same index are being moved to single node from different destinations and transfer token associated with realinstId, ensure cleanup of transferToken for realinst is done after all the transfer tokens of proxy instances.
Change-Id: I759c178b361aec485a1da0ec4939ba3b09a20719