MB-51532: jemalloc: Use Release library instead of JEMALLOC_LIBRARIES
JEMALLOC_LIBRARIES may contain a list of multiple different libraries
if there are different jemalloc library files for debug and optimised
builds.

A subsequent patch to tlm does exactly that, resulting in
JEMALLOC_LIBRARIES looking like:

    optimized;install/lib/libjemalloc.dylib;debug;install/lib/libjemallocd.dylib

instead of a single path as is does currently.

The list of build_type,path pairs is not handled by
GET_FILENAME_COMPONENT(), so change use JEMALLOC_LIBRARY_RELEASE which
only contains a single path (to the release library) - the same as
JEMALLOC_LIBRARIES used to be.

Change-Id: Ia4ee04f85b7ee13b3b310dbf854a80720021dfd7